django-site/myproject/requirement.txt
Django>=5.0
djangorestframework>=3.13
psycopg2>=2.9 